Move select all visible to multi selection model interface
It was quite well hidden in MultiSelectionModelImpl, quite hard to find.

This is an API breaking change, cannot be released in 8.0.x
