+++ server/src/main/java/com/vaadin/ui/components/grid/MultiSelectionModelImpl.java
@@ -55,39 +55,6 @@ import com.vaadin.ui.MultiSelect;
 public class MultiSelectionModelImpl<T> extends AbstractSelectionModel<T>
         implements MultiSelectionModel<T> {
 
-    /**
-     * State for showing the select all checkbox in the grid's default header
-     * row for the selection column.
-     * <p>
-     * Default value is {@link #DEFAULT}, which means that the select all is
-     * only visible if an in-memory data provider is used
-     * {@link DataProvider#isInMemory()}.
-     */
-    public enum SelectAllCheckBoxVisibility {
-        /**
-         * Shows the select all checkbox, regardless of data provider used.
-         * <p>
-         * <b>For a lazy data provider, selecting all will result in to all rows
-         * being fetched from backend to application memory!</b>
-         */
-        VISIBLE,
-        /**
-         * Never shows the select all checkbox, regardless of data provider
-         * used.
-         */
-        HIDDEN,
-        /**
-         * By default select all checkbox depends on the grid's dataprovider.
-         * <ul>
-         * <li>Visible, if the data provider is in-memory</li>
-         * <li>Hidden, if the data provider is NOT in-memory (lazy)</li>
-         * </ul>
-         *
-         * @see DataProvider#isInMemory()}.
-         */
-        DEFAULT;
-    }
-
     private class GridMultiSelectServerRpcImpl
             implements GridMultiSelectServerRpc {
 
@@ -141,17 +108,7 @@ public class MultiSelectionModelImpl<T> extends AbstractSelectionModel<T>
         return (MultiSelectionModelState) super.getState(markAsDirty);
     }
 
-    /**
-     * Sets the select all checkbox visibility mode.
-     * <p>
-     * The default value is {@link SelectAllCheckBoxVisibility#DEFAULT}, which
-     * means that the checkbox is only visible if the grid's data provider is
-     * in- memory.
-     *
-     * @param selectAllCheckBoxVisibility
-     *            the visiblity mode to use
-     * @see SelectAllCheckBoxVisibility
-     */
+    @Override
     public void setSelectAllCheckBoxVisibility(
             SelectAllCheckBoxVisibility selectAllCheckBoxVisibility) {
         if (this.selectAllCheckBoxVisibility != selectAllCheckBoxVisibility) {
@@ -160,27 +117,12 @@ public class MultiSelectionModelImpl<T> extends AbstractSelectionModel<T>
         }
     }
 
-    /**
-     * Gets the current mode for the select all checkbox visibility.
-     *
-     * @return the select all checkbox visibility mode
-     * @see SelectAllCheckBoxVisibility
-     * @see #isSelectAllCheckBoxVisible()
-     */
+    @Override
     public SelectAllCheckBoxVisibility getSelectAllCheckBoxVisibility() {
         return selectAllCheckBoxVisibility;
     }
 
-    /**
-     * Returns whether the select all checkbox will be visible with the current
-     * setting of
-     * {@link #setSelectAllCheckBoxVisibility(SelectAllCheckBoxVisibility)}.
-     *
-     * @return {@code true} if the checkbox will be visible with the current
-     *         settings
-     * @see SelectAllCheckBoxVisibility
-     * @see #setSelectAllCheckBoxVisibility(SelectAllCheckBoxVisibility)
-     */
+    @Override
     public boolean isSelectAllCheckBoxVisible() {
         updateCanSelectAll();
         return getState(false).selectAllCheckBoxVisible;
